# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RentVolt API",
    "description": "Rental market intelligence API. Licensed rental listings (RentCast), HUD Fair Market Rents, and US Census ACS — in one normalized API. Provided by **Groundwork Labs LLC**, a California limited liability company (Entity B20260059957).\n\n**Authentication:** All `/api/*` endpoints except `/api/stripe/plans`, `/api/stripe/checkout`, `/api/stripe/webhook`, `/api/stripe/session/:id`, `/api/keys/free`, `/api/feedback`, and `/api/privacy-request` require an API key in the `x-api-key` header. Paid-plan endpoints also require that the user has accepted the current Terms of Service (see `POST /api/terms/accept`).\n\n**Rate limits:** 200 requests per 15-minute window per IP globally, plus a monthly per-key quota (100 / 1,000 / 5,000 / 25,000 for Free / Growth / Scale / Enterprise respectively). When exceeded, responses return HTTP 429 with `X-RateLimit-*` headers.\n\n**Auto-renewal:** Paid plans auto-renew monthly until you cancel at `/dashboard`. Disclosure provided per California Business & Professions Code § 17602.",
    "version": "2.0.0",
    "contact": {
      "name": "Groundwork Labs LLC",
      "email": "support@groundworklabs.io",
      "url": "https://rentvolt-api.onrender.com"
    },
    "license": {
      "name": "Proprietary — see Terms of Service",
      "url": "https://rentvolt-api.onrender.com/legal/tos"
    }
  },
  "servers": [
...
  ],
...
  "paths": {
...
    "/api/scrape/listings": {
      "post": {
        "summary": "Fetch rental listings",
        "description": "Returns active long-term rental listings for a US city/state via RentCast. Supports filters (price, beds, limit, sort).",
        "tags": ["Listings"],
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListingsRequest" } } } },
        "responses": {
          "200": { "description": "Listings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListingsResponse" } } } },
          "400": { "description": "Validation error" },
          "401": { "description": "Missing API key" },
          "403": { "description": "Invalid key or terms not accepted (paid plans)" },
          "429": { "description": "Quota exceeded" }
        }
      }
    },
    "/api/scrape/bulk": {
      "post": {
        "summary": "Fetch listings for multiple cities (paid plans)",
        "description": "Up to 10 locations per request. Requires Growth, Scale, or Enterprise plan.",
        "tags": ["Listings"],
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["locations"], "properties": {
            "locations": { "type": "array", "minItems": 1, "maxItems": 10, "items": { "type": "object", "required": ["city","state"], "properties": { "city": { "type": "string" }, "state": { "type": "string" } } } },
            "filters": { "$ref": "#/components/schemas/Filters" }
          }
        } } } },
        "responses": {
          "200": { "description": "Bulk results" },
          "400": { "description": "Validation error" },
          "403": { "description": "Free plan cannot use bulk" }
        }
      }
    },
...
    "/api/market/trends/{zip}": {
      "get": {
        "summary": "Market trends for a ZIP code",
        "description": "Returns HUD Fair Market Rent + US Census ACS demographics for the given 5-digit ZIP.",
        "tags": ["Market"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "zip", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^\\d{5}$", "example": "94612" } }],
        "responses": {
          "200": { "description": "Market trends", "content": { "application/json": { "example": { "zip": "94612", "sources": ["census","hud"], "census": { "medianGrossRent": 2100, "vacancyRate": 5.6, "medianHouseholdIncome": 88000 }, "hud": { "fmr": { "twoBr": 2400 } } } } } }
        }
      }
    },
    "/api/market/briefing/{zip}": {
      "get": {
        "summary": "AI-friendly market briefing",
        "description": "One-paragraph textual briefing of a ZIP's rental market — useful for LLM-backed agents that need compact context.",
        "tags": ["Market"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "zip", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^\\d{5}$" } }],
        "responses": { "200": { "description": "Briefing text + source data" } }
      }
    },
...
    "/api/stripe/checkout": {
      "post": {
        "summary": "Start Stripe subscription checkout",
        "description": "Creates a Stripe-hosted checkout session with the CA ARL-required clear-and-conspicuous disclosures. Redirects to `/success?session_id=…` on completion.",
        "tags": ["Billing"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["plan"], "properties": { "plan": { "type": "string", "enum": ["growth","scale","enterprise"] }, "email": { "type": "string", "format": "email" } } } } } },
        "responses": {
          "200": { "description": "Checkout session", "content": { "application/json": { "example": { "sessionId": "cs_test_…", "url": "https://checkout.stripe.com/…", "plan": { "name": "RentVolt Growth", "price": "$19/mo", "monthlyRequests": 1000 } } } } },
          "400": { "description": "Invalid plan" }
        }
      }
    },
    "/api/stripe/manage": {
      "post": {
        "summary": "Open Stripe billing portal",
        "tags": ["Billing"],
        "security": [{ "ApiKeyAuth": [] }],
        "responses": { "200": { "description": "Portal URL" } }
      }
    },
    "/api/stripe/session/{id}": {
      "get": {
        "summary": "Retrieve post-checkout session (for /success page)",
        "tags": ["Billing"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Plan + key prefix + email" },
          "202": { "description": "Key still being provisioned — poll again" },
          "404": { "description": "Session not found or not paid" }
        }
      }
    },
    "/api/stripe/webhook": {
      "post": {
        "summary": "Stripe webhook handler",
        "description": "Requires valid `stripe-signature` header. Handles `checkout.session.completed`, `customer.subscription.updated`, `customer.subscription.deleted`, `invoice.payment_failed`, `invoice.paid`. Idempotent by `event.id`.",
        "tags": ["Billing"],
        "responses": { "200": { "description": "Webhook processed" }, "400": { "description": "Bad signature" }, "500": { "description": "Webhook secret not configured or handler error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "schemas": {
      "Filters": {
        "type": "object",
        "properties": {
          "maxPrice":  { "type": "integer", "minimum": 0, "maximum": 100000 },
          "minBeds":   { "type": "integer", "minimum": 0, "maximum": 20 },
          "maxBeds":   { "type": "integer", "minimum": 0, "maximum": 20 },
          "sortBy":    { "type": "string", "enum": ["price","source"], "default": "price" },
          "sortOrder": { "type": "string", "enum": ["asc","desc"], "default": "asc" },
          "limit":     { "type": "integer", "minimum": 1, "maximum": 500 }
        }
      },
      "ListingsRequest": {
        "type": "object",
        "required": ["city","state"],
        "properties": {
          "city":  { "type": "string", "example": "oakland" },
          "state": { "type": "string", "example": "ca", "minLength": 2, "maxLength": 2 },
          "filters": { "$ref": "#/components/schemas/Filters" }
        }
      },
      "Listing": {
        "type": "object",
        "properties": {
          "source":   { "type": "string", "example": "rentcast" },
          "id":       { "type": "string" },
          "price":    { "type": "integer", "nullable": true },
          "address":  { "type": "string" },
          "zip":      { "type": "string" },
          "beds":     { "type": "string" },
          "baths":    { "type": "string" },
          "sqft":     { "type": "integer", "nullable": true },
          "propertyType": { "type": "string", "nullable": true },
          "yearBuilt":{ "type": "integer", "nullable": true },
          "daysOnMarket": { "type": "integer", "nullable": true },
          "listingUrl": { "type": "string", "nullable": true },
          "latitude": { "type": "number", "nullable": true },
          "longitude":{ "type": "number", "nullable": true },
          "fetchedAt":{ "type": "string", "format": "date-time" }
        }
      },
      "ListingsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "listings": { "type": "array", "items": { "$ref": "#/components/schemas/Listing" } },
          "total": { "type": "integer" },
          "sources": { "type": "object", "additionalProperties": { "type": "integer" } },
          "query": { "type": "object" },
          "fetchedAt": { "type": "string", "format": "date-time" },
          "cached": { "type": "boolean" },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": { "type": "string" },
              "plan": { "type": "string" },
              "remaining": { "type": "integer" }
            }
          }
        }
      }
    }
  }
}